Thu, Oct 5, 2023, 12:53 AM -04:00

+++ pset_4/filter-less/helpers.c
@@ -97,7 +97,7 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            int num_pixels = 0;
+            double num_pixels = 0.0;
             int sum_red = 0;
             int sum_green = 0;
             int sum_blue = 0;
